renderer2: frameloss fixes

+++ code/renderer2/tr_scene.c
@@ -406,7 +406,7 @@ void RE_BeginScene(const refdef_t *fd)
 
 	// derived info
 
-	tr.refdef.floatTime = tr.refdef.time * 0.001f;
+	tr.refdef.floatTime = (double)tr.refdef.time * 0.001; // -EC-: cast to double
 
 	tr.refdef.numDrawSurfs = r_firstSceneDrawSurf;
 	tr.refdef.drawSurfs = backEndData->drawSurfs;
